requirements.txt
Django==1.10
django-ckeditor==5.2.1
django-modeltranslation==0.12.2
django-crispy-forms==1.6.1
django_filter==1.1.0